requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
influxdb>=5.3.0
rich>=13.0.0
elasticsearch>=7.0.0,<9.0.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
